requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
Django==1.11.13
django-auth-ldap==1.5.0
django-filter==1.1.0
djangorestframework==3.8.2
enum34==1.1.6
Fabric==1.14.0
gunicorn==19.8.1
idna==2.6
ipaddress==1.0.22
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
paramiko==2.4.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyNaCl==1.2.1
python-ldap==3.0.0
pytz==2018.4
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
